v0.0.1-FXMLupdate
An entire overhaul of the graphical system to be loaded from FXML files where each screen has a single Screen Controller. Now Engine is in the controllers package, and all Entity types are moved into the entities package leaving only App, FileOpUtils, and GameState at the root package.

App now just initiates the Stage and passes it to Engine