v0.0.1-courseMinorFix2
this time we abstract the title update to GameScreenController the onCheckTitle method now simply calls that and lets it know it's from a button press with a boolean. Since it is in GameScreenController we change how updates for onNextDay and onNextWeek work. the update() method just calls updateTitle(false) and updateAll() the abstract method